requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
pytest==8.3.3
email-validator==2.2.0

